requirements.txt
fastapi
uvicorn
spotipy
notion-client
requests
python-dotenv
